Move classic_source tests into own directory
